Add method to metadata
